requirements.txt
Flask==3.0.0
PyMySQL==1.1.0
torch==2.1.0
torchvision==0.16.0
opencv-python==4.8.1.78
numpy==1.26.0
Flask-Session==0.5.0
Werkzeug==3.0.0
